allow case of different reserved funds and accepted on sell offer update
Such a case may be caused by a bug, but in some cases it is a normal behavior
the trigger for this change is:
https://masterchain.info/selloffer.html?tx=b4fff0b33a547078cbde9a824a71a1d1ecdc6d234c0a2ac5df9123e129fa44a1&currency=MSC
Although transaction got parsed well, the automatic bug alert mechanism prevented
the sell offer from being attached to the seller address, and therefore it was
not available on the user interface, and the reserved funds did not get update.
Any user could have accepted this sell offer if they visited the sell offer
page directly.