bugfix: accept payment after sell cancel with following new sell
The following steps introduced a bug:
1. sell offer accepted
2. sell offer cancelled
3. sell accept paid within allowed time after sell cancel
4. new sell offer on thie same address

The reserved amount was temporarly negative, and non zero reserved
was treated as an open sell offer. This fact invalidated the next
new sell offer on the same address.

The relevant parsing/validating path got fixed, but the invalid
transaction stays invalid so we don't change history:
https://masterchain.info/selloffer.html?tx=13fb62038d98ca4680c6295ba10d17b63c050ccde9c4cee7579fd2e148f25581&currency=MSC
The accepts of the invalidated sell offer get also invalidated:
https://masterchain.info/sellaccept.html?tx=8593540888247a9772fbe0fbcdd765df179779ae0c728e1fe83051f1bf0efe2f&currency=MSC
https://masterchain.info/sellaccept.html?tx=67d6302a45380289de0097afdae8d21a84b0a41221ca14319b3e4cdd8952a53b&currency=MSC